Add types for a single candidate and a set of candidates.
Candidates are always an i32 (rather than a generic integer type) to
more easily perform arithmetic.  The puzzles we will deal with should
always be small enough such that we do not exceed the limits of i32.

Candidate groups are separated into an enum.  This is to help us keep
it general, as we would like to include a range option in the future.