Add unification constraint.
This constraint unifies two variables.  This is done by substituting
one variable for another in all of the current constraints, resulting
a new set of constraints.

We would like to make the equality constraint perform the more general
substitution: "coef1 var1 -> coef2 var2 + constant".  While this
constraint is less general, it can be used when the variables are not
integer quantities.